% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth2d-N.R
\name{smooth.2db}
\alias{smooth.2db}
\title{Construct a 2d-functional data object by smoothing data.}
\usage{
smooth.2db(y, Bx = NULL, By = NULL, B = NULL)
}
\arguments{
\item{y}{Tmp1}

\item{Bx}{Tmp2}

\item{By}{Tmp3}

\item{By}{Tmp4}
}
\value{
A 2d-functional object of class 'bifd'.
}
\description{
Discrete observations on one image  are fit with a set of smooth curves, each defined by an expansion in terms of user-selected basis functions.
The fitting criterion is least squares.
}
